Move some shared options methods to separate objects
