+++ modules/cli/src/main/scala/scala/cli/commands/BloopStartOptions.scala
@@ -4,6 +4,7 @@ import caseapp._
 
 import scala.build.blooprifle.BloopRifleConfig
 import scala.build.options.{BuildOptions, InternalOptions}
+import scala.cli.commands.util.JvmUtils
 
 // format: off
 final case class BloopStartOptions(
@@ -24,7 +25,7 @@ final case class BloopStartOptions(
 
   def buildOptions: BuildOptions =
     BuildOptions(
-      javaOptions = jvm.javaOptions,
+      javaOptions = JvmUtils.javaOptions(jvm),
       internal = InternalOptions(
         cache = Some(coursier.coursierCache(logging.logger.coursierLogger("")))
       )
