+++ modules/cli/src/main/scala/scala/cli/commands/Fmt.scala
@@ -5,6 +5,7 @@ import caseapp._
 import scala.build.internal.{CustomCodeWrapper, FetchExternalBinary, Runner}
 import scala.build.{CrossSources, Inputs, Logger, Sources}
 import scala.cli.CurrentParams
+import scala.cli.commands.util.SharedOptionsUtil._
 import scala.util.control.NonFatal
 
 object Fmt extends ScalaCommand[FmtOptions] {
