+++ modules/cli/src/main/scala/scala/cli/commands/FmtOptions.scala
@@ -4,6 +4,7 @@ import caseapp._
 
 import scala.build.internal.{Constants, FetchExternalBinary}
 import scala.build.options.BuildOptions
+import scala.cli.commands.util.SharedOptionsUtil._
 import scala.util.Properties
 
 // format: off
