+++ modules/cli/src/main/scala/scala/cli/commands/ScalaCliHelp.scala
