+++ modules/cli/src/main/scala/scala/cli/commands/SharedJvmOptions.scala
@@ -4,12 +4,6 @@ import caseapp._
 import com.github.plokhotnyuk.jsoniter_scala.core._
 import com.github.plokhotnyuk.jsoniter_scala.macros._
 
-import java.io.File
-
-import scala.build.options.JavaOptions
-import scala.build.{Os, Position, Positioned}
-import scala.util.Properties
-
 // format: off
 final case class SharedJvmOptions(
 
@@ -54,35 +48,8 @@ final case class SharedJvmOptions(
   @HelpMessage("Port for BSP debugging")
   @Hidden
     bspDebugPort: Option[String] = None
-) {
-  // format: on
-
-  private lazy val (javacFilePlugins, javacPluginDeps) =
-    javacPlugin
-      .filter(_.trim.nonEmpty)
-      .partition { input =>
-        input.contains(File.separator) ||
-        (Properties.isWin && input.contains("/")) ||
-        input.count(_ == ':') < 2
-      }
-
-  def javaOptions = JavaOptions(
-    javaHomeOpt = javaHome.filter(_.nonEmpty).map(v =>
-      Positioned(Seq(Position.CommandLine("--java-home")), os.Path(v, Os.pwd))
-    ),
-    jvmIdOpt = jvm.filter(_.nonEmpty),
-    jvmIndexOpt = jvmIndex.filter(_.nonEmpty),
-    jvmIndexOs = jvmIndexOs.map(_.trim).filter(_.nonEmpty),
-    jvmIndexArch = jvmIndexArch.map(_.trim).filter(_.nonEmpty),
-    javacPluginDependencies = SharedOptions.parseDependencies(
-      javacPluginDeps.map(Positioned.none(_)),
-      ignoreErrors = false
-    ),
-    javacPlugins = javacFilePlugins.map(s => Positioned.none(os.Path(s, Os.pwd))),
-    javacOptions = javacOption
-  )
-
-}
+)
+// format: on
 
 object SharedJvmOptions {
   lazy val parser: Parser[SharedJvmOptions]                           = Parser.derive
