+++ modules/cli/src/main/scala/scala/cli/commands/util/Cached.scala
@@ -1,0 +1,7 @@
+package scala.cli.commands.util
+
+object cached {
+  private val cache = collection.concurrent.TrieMap[Any, Any]()
+
+  def apply[V](from: Any)(create: => V): V = cache.getOrElseUpdate(from, create).asInstanceOf[V]
+}
